requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
cryptography==42.0.0
django-filter==24.1
requests==2.31.0
